cloud-functions/linkedin_scraper/requirements.txt
google-api-python-client==2.149.0
linkedin-scraper==2.9.0
selenium==3.141.0
urllib3<2
functions-framework==3.0.0
webdriver-manager
google-cloud-storage==2.18.2
gunicorn==20.1.0
